requirement.txt
 Streamlit
 google-generativeai
 python-dotenv
 pdf2image
 